// src/api/enums/TransactionType.ts
export enum TransactionType {
    FUNDING = "funding",
    TRANSFER = "transfer",
    WITHDRAWAL = "withdrawal",
    WALLET_TRANSFER = "wallet_transfer",
    PAYMENT = "payment",
  }

export enum TransactionTypeAccount {
    FUNDING = "001",
    TRANSFER = "004",
    WITHDRAWAL = "002",
    WALLET_TRANSFER = "008",
    PAYMENT = "016",
  }